Reject zero-byte continuation fragments
Reject empty continuation payloads in `InboundSeriesTracker` so an
attacker cannot keep a fragmented series alive without consuming any of
its remaining body budget.

Add a regression test that exercises the tracker-level guard directly.
The codec test module moves into `frame_tests.rs` so `frame.rs` stays
below the repository's module-size lint limit.
